backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
python-dotenv==1.0.0
httpx==0.26.0
python-jose==3.3.0
zoomus==1.2.1
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
openai==1.12.0
pydantic==1.10.16
typing-extensions==4.7.1
python-multipart==0.0.6
websockets==12.0
PyJWT==2.8.0
aiohttp==3.9.1
deepgram-sdk==2.12.0
